manage listing by user and check user ownership

+++ app/Http/Controllers/ListingController.php
@@ -60,6 +60,12 @@ class ListingController extends Controller
 
     //Update listing data
     public function update(Request $request, Listing $listing){
+
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -84,8 +90,18 @@ class ListingController extends Controller
     //delete listing
     public function destroy(Listing $listing)
     {
-        # code...
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message','Listing deleted successfully');
+    }
+
+    //manage listing function
+    public function manage()
+    {
+        return view('listings.manage',['listings' => auth()->user()->listings()->get()]);
     }
 }
